Drop int8

+++ src/Calvano2020/policy.jl
@@ -21,7 +21,7 @@ CalvanoPolicy(env::CalvanoEnv) = MultiAgentManager(
             ),
             trajectory = VectorSARTTrajectory(;
                 state = Int16,
-                action = Union{Int8,NoOp},
+                action = Union{Int,NoOp},
                 reward = Float32,
                 terminal = Bool,
             ),
